data table JS bugfix
after changing an option in the cog-menu (exclude low pop or flatten), the number of rows in the report changes. JS used to pass a totalRows parameter to the server that prevented it from recalculating the number of rows. fix: when an option in the cog-menu is changed, delete the totalRows parameter so that PHP counts the number of rows.